backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
pydantic==1.10.7
python-multipart==0.0.6
httpx==0.23.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.9
alembic==1.10.3
youtube-transcript-api==0.6.1
youtube-search-python==1.6.6
openai==0.27.4
tiktoken>=0.4.0
